Follow gsctl convention from Command-Blueprint.
